Fix Postgres migration... again.

+++ Arrowgene.Ddon.Database/Files/Database/Script/schema_sqlite.sql
@@ -927,8 +927,8 @@ CREATE TABLE IF NOT EXISTS "ddon_equipment_limit_break"
     "item_uid"         TEXT       NOT NULL,
     "effect_id"        INTEGER    NOT NULL,
     "unk1"             INTEGER    NOT NULL,
-    "effect_type"      TINYINT    NOT NULL,
-    "unk0"             TINYINT    NOT NULL,
+    "effect_type"      SMALLINT    NOT NULL,
+    "unk0"             SMALLINT    NOT NULL,
     CONSTRAINT "pk_ddon_equipment_limit_break" PRIMARY KEY ("character_id", "item_uid", "effect_type"),
     CONSTRAINT "fk_ddon_equipment_limit_break_item_uid" FOREIGN KEY ("item_uid") REFERENCES "ddon_storage_item" ("item_uid") ON DELETE CASCADE,
     CONSTRAINT "fk_ddon_equipment_limit_break_character_id" FOREIGN KEY ("character_id") REFERENCES "ddon_character" ("character_id") ON DELETE CASCADE
